+++ xarray/backends/file_manager.py
@@ -11,8 +11,8 @@ from typing import Any
 
 from xarray.backends.locks import acquire
 from xarray.backends.lru_cache import LRUCache
-from xarray.core import utils
 from xarray.core.options import OPTIONS
+from xarray.namedarray import utils
 
 # Global cache for storing open files.
 FILE_CACHE: LRUCache[Any, io.IOBase] = LRUCache(
