+++ xarray/core/coordinates.py
@@ -26,11 +26,11 @@ from xarray.core.merge import merge_coordinates_without_align, merge_coords
 from xarray.core.types import DataVars, Self, T_DataArray, T_Xarray
 from xarray.core.utils import (
     Frozen,
-    ReprObject,
     either_dict_or_kwargs,
     emit_user_level_warning,
 )
 from xarray.core.variable import Variable, as_variable, calculate_dimensions
+from xarray.namedarray.utils import ReprObject
 
 if TYPE_CHECKING:
     from xarray.core.common import DataWithCoords
